♻️ refactor: consolidate model bank provider utilities
