+++ packages/model-bank/src/aiModels/index.ts
@@ -212,6 +212,7 @@ export const loadModels = async (
   return buildDefaultModelList(modelMap);
 };
 
+export { gptImage1Schema, gptImage2Schema } from '../const/imageParameters';
 export { default as ai21 } from './ai21';
 export { default as ai302 } from './ai302';
 export { default as ai360 } from './ai360';
@@ -259,7 +260,6 @@ export { default as nvidia } from './nvidia';
 export { default as ollama } from './ollama';
 export { default as ollamacloud } from './ollamacloud';
 export { default as openai, openaiChatModels } from './openai';
-export { gptImage1Schema, gptImage2Schema } from './openaiImageParameters';
 export { default as opencodecodingplan } from './opencodeCodingPlan';
 export { default as opencodezen } from './opencodeZen';
 export { default as openrouter } from './openrouter';
