+++ packages/model-bank/src/index.ts
@@ -2,3 +2,4 @@ export * from './aiModels';
 export * from './const/modelProvider';
 export * from './standard-parameters';
 export * from './types';
+export * from './utils';
