+++ src/server/routers/lambda/__tests__/video.test.ts
@@ -9,7 +9,7 @@ import { AsyncTaskStatus } from '@/types/asyncTask';
 const {
   mockAfter,
   mockCreateVideo,
-  mockIsLobeHubModelAvailable,
+  mockLoadModels,
   mockProcessBackgroundVideoPolling,
   mockResolveBusinessModelMapping,
   mockServerDB,
@@ -19,13 +19,13 @@ const {
   const mockServerDB = { transaction: mockTransaction };
   const mockCreateVideo = vi.fn();
   const mockAfter = vi.fn((cb: () => void) => cb());
-  const mockIsLobeHubModelAvailable = vi.fn();
+  const mockLoadModels = vi.fn();
   const mockProcessBackgroundVideoPolling = vi.fn().mockResolvedValue(undefined);
   const mockResolveBusinessModelMapping = vi.fn();
   return {
     mockAfter,
     mockCreateVideo,
-    mockIsLobeHubModelAvailable,
+    mockLoadModels,
     mockProcessBackgroundVideoPolling,
     mockResolveBusinessModelMapping,
     mockServerDB,
@@ -59,7 +59,7 @@ vi.mock('@lobechat/business-model-runtime', async (importOriginal) => ({
     mockResolveBusinessModelMapping(...args),
 }));
 vi.mock('@lobechat/business-model-bank/model-config', () => ({
-  isLobeHubModelAvailable: (...args: [string, string]) => mockIsLobeHubModelAvailable(...args),
+  loadModels: mockLoadModels,
 }));
 vi.mock('@/business/server/video-generation/getVideoFreeQuota', () => ({
   getVideoFreeQuota: vi.fn().mockResolvedValue({ remaining: 10 }),
@@ -146,7 +146,15 @@ describe('videoRouter', () => {
         resolvedModelId: model,
       }),
     );
-    mockIsLobeHubModelAvailable.mockResolvedValue(true);
+    mockLoadModels.mockResolvedValue([
+      {
+        abilities: {},
+        enabled: true,
+        id: 'dreamina-seedance-2-0-260128',
+        providerId: 'lobehub',
+        type: 'video',
+      },
+    ]);
   });
 
   describe('createVideo - async strategy routing', () => {
