+++ src/server/routers/lambda/image/index.test.ts
@@ -12,7 +12,7 @@ const {
   mockAsyncTaskModelUpdate,
   mockChargeBeforeGenerate,
   mockCreateAsyncCaller,
-  mockIsLobeHubModelAvailable,
+  mockLoadModels,
   mockResolveBusinessModelMapping,
 } = vi.hoisted(() => ({
   mockServerDB: {
@@ -23,7 +23,7 @@ const {
   mockAsyncTaskModelUpdate: vi.fn(),
   mockChargeBeforeGenerate: vi.fn(),
   mockCreateAsyncCaller: vi.fn(),
-  mockIsLobeHubModelAvailable: vi.fn(),
+  mockLoadModels: vi.fn(),
   mockResolveBusinessModelMapping: vi.fn(),
 }));
 
@@ -64,7 +64,7 @@ vi.mock('@lobechat/business-model-runtime', async (importOriginal) => ({
 }));
 
 vi.mock('@lobechat/business-model-bank/model-config', () => ({
-  isLobeHubModelAvailable: (...args: [string, string]) => mockIsLobeHubModelAvailable(...args),
+  loadModels: mockLoadModels,
 }));
 
 // Mock async caller
@@ -124,7 +124,15 @@ describe('imageRouter', () => {
     mockChargeBeforeGenerate.mockResolvedValue(undefined);
     mockGetKeyFromFullUrl.mockResolvedValue(null);
     mockGetFullFileUrl.mockResolvedValue(null);
-    mockIsLobeHubModelAvailable.mockResolvedValue(true);
+    mockLoadModels.mockResolvedValue([
+      {
+        abilities: {},
+        enabled: true,
+        id: 'gpt-image-1',
+        providerId: 'lobehub',
+        type: 'image',
+      },
+    ]);
 
     // Setup default transaction mock
     const mockBatch = {
