+++ src/server/routers/lambda/image/index.ts
@@ -1,10 +1,11 @@
 import { BRANDING_PROVIDER } from '@lobechat/business-const';
-import { isLobeHubModelAvailable } from '@lobechat/business-model-bank/model-config';
+import { loadModels } from '@lobechat/business-model-bank/model-config';
 import { resolveBusinessModelMapping } from '@lobechat/business-model-runtime';
 import { ChatErrorType } from '@lobechat/types';
 import { TRPCError } from '@trpc/server';
 import debug from 'debug';
 import { and, eq } from 'drizzle-orm';
+import { isProviderModelAvailable } from 'model-bank';
 import { z } from 'zod';
 
 import { chargeBeforeGenerate } from '@/business/server/image-generation/chargeBeforeGenerate';
@@ -71,7 +72,7 @@ export const imageRouter = router({
     // can't serve the requested id.
     if (
       provider === BRANDING_PROVIDER &&
-      !(await isLobeHubModelAvailable(resolvedModelId, 'image'))
+      !isProviderModelAvailable(await loadModels(), BRANDING_PROVIDER, resolvedModelId, 'image')
     ) {
       throw new TRPCError({
         cause: { data: { modelType: 'image', requestedModel: model } },
